+++ Changes.md
@@ -7,4 +7,8 @@ Version next
 * Add: Allows sitecore.tools to be added to the PATH environment variable (2015-11-01) - Martin Svarrer Christensen
 * Add: Troubleshooting task; republish, rebuild search indexes, rebuild link database (2015-11-01) - SPEAK team
 * Add: Revamped the entire process for creating a new project (2015-11-04) - Martin Svarrer Christensen
-* Add: reset-website for deleting items and files from the website (2015-11-06) - SPEAK team
+* Add: reset-website tak for deleting items and files from the website (2015-11-06) - SPEAK team
+* Add: install-project task for installing a project directly from the project directory (2015-11-06)
+* Add: watch-project task for watching a project for changes and installing the project (2015-11-06)
+* Add: build-project:force-update setting which indicates if media and files are always overwritten (2015-11-06)
+* Add: build-project:file-search-pattern sets the file search pattern for project directory visitor (2015-11-06)